feat: [#186333254] remove unnecessary props in sidabar cards
