+++ content/src/display-content/roadmap-sidebar-cards/formation-nudge.md
@@ -3,9 +3,7 @@ id: formation-nudge
 header: Find Your Deadlines and Opportunities
 hasCloseButton: false
 ctaText: See My Deadlines and Opportunities
-weight: 1
 preBodySpanButtonText: ""
-section: below-opportunities
 ---
 
 If you [formed your business](/tasks/form-business-entity), you can see:
