+++ content/src/display-content/roadmap-sidebar-cards/funding-nudge.md
@@ -3,9 +3,7 @@ id: funding-nudge
 header: Need Funding?
 hasCloseButton: false
 ctaText: Show Me Funding Opportunities
-weight: 2
 preBodySpanButtonText: ""
-section: above-opportunities
 ---
 
 If you're done setting up your business, we can hide your task list and show you possible funding opportunities.
