+++ content/src/display-content/roadmap-sidebar-cards/go-to-profile-nudge.md
@@ -3,9 +3,7 @@ id: go-to-profile
 header: Customize Your Experience
 hasCloseButton: false
 ctaText: ""
-weight: 3
 preBodySpanButtonText: Go to your profile
-section: "above-opportunities"
 ---
 
 to customize your certifications, funding opportunities, and annual report deadline.
