+++ content/src/display-content/roadmap-sidebar-cards/not-registered-existing-account.md
@@ -3,9 +3,7 @@ id: not-registered-existing-account
 header: "Link Your myNewJersey Account"
 hasCloseButton: false
 ctaText: ""
-weight: 1
 preBodySpanButtonText: ""
-section: above-opportunities
 ---
 
 To save your progress on your Business.NJ.gov account, [link your existing myNewJersey account](/account-setup).
