+++ content/src/display-content/roadmap-sidebar-cards/not-registered.md
@@ -4,8 +4,6 @@ header: "Link with myNewJersey to Save Your Progress"
 hasCloseButton: false
 ctaText: ""
 preBodySpanButtonText: ""
-weight: 1
-section: above-opportunities
 ---
 
 Now that you've filled out your business information, you can save your progress and [link your Business.NJ.gov account with myNewJersey](/account-setup).
