+++ web/src/components/dashboard/SidebarCardsContainer.tsx
@@ -1,13 +1,20 @@
 import { SidebarCardsList } from "@/components/dashboard/SidebarCardsList";
 import { useUserData } from "@/lib/data-hooks/useUserData";
-import { filterCertifications } from "@/lib/domain-logic/filterCertifications";
-import { filterFundings } from "@/lib/domain-logic/filterFundings";
-import { getVisibleCertifications } from "@/lib/domain-logic/getVisibleCertifications";
-import { getVisibleFundings } from "@/lib/domain-logic/getVisibleFundings";
-import { sortCertifications } from "@/lib/domain-logic/sortCertifications";
-import { sortFundings } from "@/lib/domain-logic/sortFundings";
+import {
+  filterCertifications,
+  filterFundings,
+  getForYouCardCount,
+  getHiddenCertifications,
+  getHiddenFundings,
+  getVisibleCertifications,
+  getVisibleFundings,
+  getVisibleSideBarCards,
+  sortCertifications,
+  sortFundings,
+} from "@/lib/domain-logic/sidebarCardsHelpers";
 import { Certification, Funding, SidebarCardContent } from "@/lib/types/types";
-import { Business, LookupOperatingPhaseById } from "@businessnjgovnavigator/shared";
+import { LookupOperatingPhaseById } from "@businessnjgovnavigator/shared/";
+import { isRemoteWorkerOrSellerBusiness } from "@businessnjgovnavigator/shared/domain-logic/businessPersonaHelpers";
 import { ReactElement } from "react";
 
 interface Props {
@@ -19,83 +26,35 @@ interface Props {
 export const SidebarCardsContainer = (props: Props): ReactElement => {
   const { business } = useUserData();
 
-  const filteredSortedFundings = business ? sortFundings(filterFundings(props.fundings, business)) : [];
-
-  const filteredSortedCertifications = business
-    ? sortCertifications(filterCertifications(props.certifications, business))
-    : [];
-
-  const visibleSortedFundings = getVisibleFundings(filteredSortedFundings, business as Business);
-
+  const visibleSidebarCards = getVisibleSideBarCards(business, props.sidebarDisplayContent);
+  const visibleSortedFundings = getVisibleFundings(
+    sortFundings(filterFundings({ fundings: props.fundings, business: business })),
+    business
+  );
+  const hiddenSortedFundings = sortFundings(getHiddenFundings(business, props.fundings));
   const visibleSortedCertifications = getVisibleCertifications(
-    filteredSortedCertifications,
-    business as Business
+    sortCertifications(filterCertifications({ certifications: props.certifications, business: business })),
+    business
   );
-
   const hiddenSortedCertifications = sortCertifications(
-    (business?.preferences.hiddenCertificationIds || [])
-      .map((id) => {
-        return props.certifications.find((it) => {
-          return it.id === id;
-        });
-      })
-      .filter((it) => {
-        return it !== undefined;
-      }) as Certification[]
+    getHiddenCertifications(business, props.certifications)
   );
 
-  const hiddenSortedFundings = sortFundings(
-    (business?.preferences.hiddenFundingIds || [])
-      .map((id) => {
-        return props.fundings.find((it) => {
-          return it.id === id;
-        });
-      })
-      .filter((it) => {
-        return it !== undefined;
-      }) as Funding[]
-  );
-
-  const displayFundingCards = (): boolean => {
-    return LookupOperatingPhaseById(business?.profileData.operatingPhase).displayFundings;
-  };
-
-  const displayCertificationsCards = (): boolean => {
-    return LookupOperatingPhaseById(business?.profileData.operatingPhase).displayCertifications;
-  };
-
-  const visibleCardsOrderedByWeight = business
-    ? business.preferences.visibleSidebarCards
-        .map((id: string) => {
-          return props.sidebarDisplayContent[id];
-        })
-        .sort((cardA: SidebarCardContent, cardB: SidebarCardContent): number => {
-          return cardA.weight < cardB.weight ? -1 : 1;
-        })
-    : [];
-
-  const getTopCards = (): SidebarCardContent[] => {
-    return visibleCardsOrderedByWeight.filter((card) => {
-      return card.section === "above-opportunities";
-    });
-  };
-
-  const getBottomCards = (): SidebarCardContent[] => {
-    return visibleCardsOrderedByWeight.filter((card) => {
-      return card.section === "below-opportunities";
-    });
-  };
+  const remoteSellerWorker = isRemoteWorkerOrSellerBusiness(business);
 
   return (
     <SidebarCardsList
-      topCards={getTopCards()}
-      bottomCards={getBottomCards()}
+      sideBarCards={visibleSidebarCards}
       fundings={visibleSortedFundings}
       hiddenFundings={hiddenSortedFundings}
       certifications={visibleSortedCertifications}
       hiddenCertifications={hiddenSortedCertifications}
-      displayFundings={displayFundingCards()}
-      displayCertifications={displayCertificationsCards()}
+      isRemoteSellerWorker={remoteSellerWorker}
+      displayFundingCards={LookupOperatingPhaseById(business?.profileData.operatingPhase).displayFundings}
+      displayCertificationsCards={
+        LookupOperatingPhaseById(business?.profileData.operatingPhase).displayCertifications
+      }
+      cardCount={getForYouCardCount(business, props.certifications, props.fundings)}
     />
   );
 };
