--- web/src/lib/domain-logic/getVisibleFundings.ts
@@ -1,8 +1,0 @@
-import { Business } from "@businessnjgovnavigator/shared";
-import { Funding } from "../types/types";
-
-export const getVisibleFundings = (fundings: Funding[], business: Business): Funding[] => {
-  return fundings.filter((it) => {
-    return !business?.preferences.hiddenFundingIds.includes(it.id);
-  });
-};
